refactor(export): gộp export query vào print query, validate loai qua Create()
